requirements.txt
Flask
onnxruntime
Pillow
matplotlib
seaborn
numpy